Add files via upload
SPEAKER_00 [0.00–3.00]:  We missed the top line of revenue and numbers.
SPEAKER_00 [3.00–13.00]:  So, I just wanted to tell you that we will be looking at the overall revenue and the numbers of the pad in the future presentations.
SPEAKER_00 [13.00–15.00]:  This is the format.
SPEAKER_01 [15.00–27.00]:  So, any questions, anything in and quite honestly, if whatever people need to understand, if they want to improve these presentations in whichever way or form,
SPEAKER_01 [27.00–33.00]:  please, this is a good time to have that discussion. Anybody would like to be able to.
SPEAKER_01 [33.00–37.00]:  Kumar?
SPEAKER_01 [37.00–40.00]:  One thing is important. We miss one thing.
SPEAKER_00 [40.00–46.00]:  So, what we can do is keep in mind customer acquisition.
SPEAKER_00 [46.00–56.00]:  How many customers added? Not about links, not about any other deployments, but how many new customers, what is the upsell?
SPEAKER_00 [56.00–58.00]:  How much is the upsell of an existing customer?
SPEAKER_00 [58.00–60.00]:  What is your revenue contribution?